add NextMoveTactic
This uses stockfish to find a suitable next for "self"  move
when proving a `ForcedWin`.  It assumes that a binary called
`stockfish` is in your path.

I modeled it after the `polyrith` tactic, so you can click on
`try this: move "foo"` and it will insert the suggestions.
Right now it let's stockfish think for 0.1s,
but we could make that adjustable of course.